feat: Expand version format to four components
This expands our version format to four components and retains leading
zeros in formatting.  The intention is to support more flexible version
formats some of which are common in the Microsoft environment where
versions are in the form `major.minor.build.revision`.  We map these to
`major.minor.patch.revision`.