fix: Return borrowed lifetime from all accessor methods (#43)
The data in Release and Version is borrowed from the parsed string, but the
lifetimes of all the component methods (e.g. build_code) was tied to the
parsed instance. This makes it inconvenient to use the release parser, as the
parsed instance has to be held for longer, and it cannot be used in temporaries.